refactor: add Query prefix to hooks
