+++ src/components/wallet/WalletJarsDetailsContent.tsx
@@ -4,7 +4,7 @@ import { AlertTriangleIcon } from 'lucide-react'
 import { Trans, useTranslation } from 'react-i18next'
 import { useStore } from 'zustand'
 import { useJars, type Jar } from '@/context/JamWalletInfoContext'
-import type { Utxo } from '@/hooks/useUtxos'
+import type { Utxo } from '@/hooks/useQueryUtxos'
 import { cn } from '@/lib/utils'
 import { jamSettingsStore } from '@/store/jamSettingsStore'
 import type { JarIndex } from '@/types/global'
